requirements.txt
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.1
idna==3.4
pycparser==2.21
PyJWT==2.6.0
requests==2.28.2
urllib3==1.26.15
